+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+- Lemmas: `fset_eqP` `fset_eq0E` `domm_filter_eq` `fperm1V`
+
 ## Changed
 
 ## Deprecated